Namespace code changes to match the Misc folder renaming to Common
