+++ XamlSamples/XamlSamples.sln
@@ -238,7 +238,6 @@ Global
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.AppStore|x64.ActiveCfg = AppStore|iPhone
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.AppStore|x86.ActiveCfg = AppStore|iPhone
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.Debug|Any CPU.ActiveCfg = Debug|iPhoneSimulator
-		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.Debug|Any CPU.Build.0 = Debug|iPhoneSimulator
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.Debug|ARM.ActiveCfg = Debug|iPhone
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.Debug|iPhone.ActiveCfg = Debug|iPhone
 		{9C9E6DD6-3CE5-4815-A85A-4953AD5509D2}.Debug|iPhone.Build.0 = Debug|iPhone
