+++ XamlSamples/XamlSamples/XamlSamples/Views/StaticConstantsPage.xaml
@@ -1,22 +1,22 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:misc="clr-namespace:XamlSamples.Misc"
+             xmlns:common="using:XamlSamples.Common"
              xmlns:sys="clr-namespace:System;assembly=mscorlib"
              x:Class="XamlSamples.Views.StaticConstantsPage"
              Title="Static Constants Page"
-             Padding="{x:Static misc:AppConstants.PagePadding}">
+             Padding="{x:Static common:AppConstants.PagePadding}">
 
     <StackLayout>
         <Label Text="Hello, XAML!"
-               TextColor="{x:Static misc:AppConstants.BackgroundColor}"
-               BackgroundColor="{x:Static misc:AppConstants.ForegroundColor}"
-               Font="{x:Static misc:AppConstants.TitleFont}"
+               TextColor="{x:Static common:AppConstants.BackgroundColor}"
+               BackgroundColor="{x:Static common:AppConstants.ForegroundColor}"
+               Font="{x:Static common:AppConstants.TitleFont}"
                HorizontalOptions="Center" />
 
         <BoxView WidthRequest="{x:Static sys:Math.PI}"
                  HeightRequest="{x:Static sys:Math.E}"
-                 Color="{x:Static misc:AppConstants.ForegroundColor}"
+                 Color="{x:Static common:AppConstants.ForegroundColor}"
                  HorizontalOptions="Center"
                  VerticalOptions="CenterAndExpand"
                  Scale="100" />
